Added parameter to use source objects instead of compound path
