+++ package.json
@@ -100,6 +100,8 @@
       ".ts"
     ],
     "reporter": [
+      "text",
+      "text-summary",
       "lcovonly"
     ]
   },
