+++ src/cli/check-argv.ts
@@ -12,9 +12,10 @@ export function checkArgv(): void {
 
   console.log(`${logger.bold('Usage:')} xd2svg-cli InputFile.xd [options]
   ${logger.bold('options:')}
-  ${logger.bold('-o, --output')}       - specify output path (default FileName directory or FileName.svg)
-  ${logger.bold('-s, --single')}       - specify does output should be single file with all artboards or directory with separated each other (default: false)
-  ${logger.bold('-p, --pretty-print')} - specify does output should be pretty printed (default: false)
+  ${logger.bold('-o, --output')}           - output path (default FileName directory or FileName.svg)
+  ${logger.bold('-s, --single')}           - single file output (default: false)
+  ${logger.bold('-p, --pretty-print')}     - pretty printed output (default: false)
+  ${logger.bold('--prefer-compound-path')} - use compound path instead of source objects (default: true)
 
   For additional information: man xd2svg-cli`);
