+++ src/core/artboard-converter.ts
@@ -6,20 +6,18 @@
  * found in the LICENSE file at https://github.com/L2jLiga/xd2svg/LICENSE
  */
 
-import * as builder                             from 'xmlbuilder';
-import { XMLElementOrXMLNode }                  from 'xmlbuilder';
-import * as logger                              from '../utils/logger';
-import { createShape, TextConverter }           from './converters';
-import { createStyles }                         from './create-styles';
-import { Artboard, ArtboardInfo, Shape }        from './models';
-import { applyIfPossible }                      from './utils';
-import { SvgRectToClipPath, svgRectToClipPath } from './utils/svg-rect-to-clip-path';
+import * as builder               from 'xmlbuilder';
+import { Options }                from '../common';
+import { createStyles }           from './create-styles';
+import { Artboard, ArtboardInfo } from './models';
+import { applyIfPossible }        from './utils';
+import { createElem }             from './utils/create-elem';
 
-export function artboardConverter(artboardsRoot: Artboard, artboardInfo: ArtboardInfo, prettyPrint?: boolean): string[] {
-  return artboardsRoot.children.map(toArtboards(artboardInfo, prettyPrint));
+export function artboardConverter(artboardsRoot: Artboard, artboardInfo: ArtboardInfo, options: Options): string[] {
+  return artboardsRoot.children.map(toArtboards(artboardInfo, options));
 }
 
-function toArtboards(artboardInfo: ArtboardInfo, pretty: boolean) {
+function toArtboards(artboardInfo: ArtboardInfo, options: Options) {
   return (root: Artboard): string => {
     const svg = builder.begin().element('svg', {
       'enable-background': `new ${artboardInfo.x} ${artboardInfo.y} ${artboardInfo.width} ${artboardInfo.height}`,
@@ -42,89 +40,6 @@ function toArtboards(artboardInfo: ArtboardInfo, pretty: boolean) {
 
     const defs = svg.element('defs');
 
-    return createElem(root.artboard, svg, defs).end({pretty});
+    return createElem(root.artboard, svg, defs, options).end({pretty: options.prettyPrint});
   };
-}
-
-function createTransforms(src): string {
-  return `matrix(${src.a}, ${src.b}, ${src.c}, ${src.d}, ${src.tx}, ${src.ty})`;
-}
-
-export function createElem(svgObjCollection: Artboard, parentElement: XMLElementOrXMLNode, defs: XMLElementOrXMLNode): XMLElementOrXMLNode {
-  svgObjCollection.children
-    .map((svgObject: Artboard): void => {
-      let node: XMLElementOrXMLNode;
-
-      switch (svgObject.type) {
-        case 'shape':
-          const shape = svgObject.shape;
-          node = createShape(svgObject.shape, parentElement);
-          applyBorderRadius(svgObject, shape, defs);
-          break;
-
-        case 'text':
-          node = TextConverter.createText(svgObject.text, parentElement).result;
-          break;
-
-        case 'group':
-          node = parentElement.element('g');
-          createElem(svgObject.group, node, defs);
-          break;
-
-        default:
-          console.warn(`${logger.bold(logger.red('Create element:'))} unknown type: %j`, svgObject);
-          return;
-      }
-
-      applyAttributes(node, defs, svgObject);
-    });
-
-  return parentElement;
-}
-
-function applyBorderRadius(svgObject: Artboard, shape: Shape, defs: XMLElementOrXMLNode) {
-  if (shape.type === 'rect' && shape.r) {
-    const {width, height, r} = shape;
-    const ref = `clip-path-${width}-${height}-${r.join('-')}`;
-
-    createClipPath({width, height, r}, ref, defs);
-
-    svgObject.style = svgObject.style || {};
-    svgObject.style.clipPath = {ref};
-  }
-}
-
-function createClipPath(data: SvgRectToClipPath, id: string, defs: XMLElementOrXMLNode) {
-  const {width, height, r} = data;
-  const maxBR = Math.min(width, height) / 2;
-
-  const clipPath = svgRectToClipPath({
-    height,
-    r: [
-      Math.min(r[0], maxBR),
-      Math.min(r[1], maxBR),
-      Math.min(r[2], maxBR),
-      Math.min(r[3], maxBR),
-    ],
-    width,
-  }, defs);
-
-  clipPath.attribute('id', id);
-}
-
-function applyAttributes(node: XMLElementOrXMLNode, defs: XMLElementOrXMLNode, svgObject: Artboard) {
-  applyIfPossible(node, 'id', svgObject.id);
-  applyIfPossible(node, 'name', svgObject.name);
-
-  if (svgObject.visible === false) {
-    svgObject.style ? svgObject.style.display = 'none' : svgObject.style = {display: 'none'};
-  }
-
-  if (svgObject.style) {
-    node.attribute('style', createStyles(svgObject.style, defs));
-  }
-
-  if (svgObject.transform) {
-    node.attribute('transform', createTransforms(svgObject.transform));
-  }
 }
