+++ src/xd2svg.ts
@@ -6,27 +6,19 @@
  * found in the LICENSE file at https://github.com/L2jLiga/xd2svg/LICENSE
  */
 
-import * as extractZip                                                  from 'extract-zip';
-import { existsSync, lstatSync, writeFileSync }                         from 'fs';
-import { dirSync, SynchrounousResult }                                  from 'tmp';
-import { promisify }                                                    from 'util';
-import { defaultOptions, Dictionary, Directory, Options, OutputFormat } from './common';
-import { proceedFile }                                                  from './core';
-import * as logger                                                      from './utils/logger';
+import * as extractZip                                                                                from 'extract-zip';
+import { existsSync, lstatSync, writeFileSync }                                                       from 'fs';
+import { dirSync, SynchrounousResult }                                                                from 'tmp';
+import { promisify }                                                                                  from 'util';
+import { defaultOptions, Dictionary, Directory, MultipleOutput, Options, OutputFormat, SingleOutput } from './common';
+import { proceedFile }                                                                                from './core';
+import * as logger                                                                                    from './utils/logger';
 
 const extract = promisify(extractZip);
 
-interface SingleOutput extends Options {
-  single: true;
-}
-
-interface MultipleOutput extends Options {
-  single: false;
-}
-
-export default async function xd2svg(input: string | Buffer, options: SingleOutput): Promise<string>;
+export default async function xd2svg(input: string | Buffer, options?: SingleOutput): Promise<string>;
 export default async function xd2svg(input: string | Buffer, options?: MultipleOutput): Promise<Dictionary<string>>;
-export default async function xd2svg(input: string | Buffer, options: Options): Promise<OutputFormat>;
+export default async function xd2svg(input: string | Buffer, options?: Options): Promise<OutputFormat>;
 export default async function xd2svg(input: string | Buffer, options: Options = defaultOptions): Promise<OutputFormat> {
   const opts: Options = {
     ...defaultOptions,
@@ -34,7 +26,7 @@ export default async function xd2svg(input: string | Buffer, options: Options = 
   };
 
   const directory: Directory = await openMockup(input);
-  const svg: string | Dictionary<string> = proceedFile(directory, opts.single, opts.prettyPrint);
+  const svg: string | Dictionary<string> = proceedFile(directory, opts);
 
   if (directory.removeCallback) directory.removeCallback();
